% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_stan.R
\name{glm_binomial}
\alias{glm_binomial}
\title{Bayesian binomial regression with Stan}
\usage{
glm_binomial(X, y, n, link = 0, mu0 = rep(0, ncol(X)),
  Sigma0 = diag(1, ncol(X)), method = "sampling", ...)
}
\arguments{
\item{X}{Numeric matrix of design.}

\item{y}{Numeric vector of responses.}

\item{n}{Numeric vector of number of trials.}

\item{mu0}{Prior mean for regression coefficients.}

\item{Sigma0}{Prior covariance for regression coefficients.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian binomial regression with Stan
}
