% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalisation_stan.R
\name{lassoreg_stan}
\alias{lassoreg_stan}
\title{Bayesian lasso regression with Stan}
\usage{
lassoreg_stan(X, y, ...)
}
\arguments{
\item{X}{Design matrix (no intercept)}

\item{y}{Response vector}

\item{...}{Additional arguments}
}
\value{
An object of class `stanfit`
}
\description{
Bayesian lasso regression with Stan
}
