% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg_stan.R
\name{logreg_stan_pred}
\alias{logreg_stan_pred}
\title{Bayesian logistic regression with Stan}
\usage{
logreg_stan_pred(X, y, n, Xpred, npred, ...)
}
\arguments{
\item{X}{Numeric matrix of design.}

\item{y}{Numeric vector of responses.}

\item{n}{Numeric vector of number of trials.}

\item{Xpred}{Numeric matrix of prediction.}

\item{npred}{Numeric vector of number of trials to predict.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian logistic regression with Stan
}
