% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survreg_stan.R
\name{survreg_ph_exponential}
\alias{survreg_ph_exponential}
\title{Bayesian proportional hazards exponential model}
\usage{
survreg_ph_exponential(X, y, v, rcens = rep(0, length(y)), mu0 = rep(0,
  ncol(X)), Sigma0 = diag(1, ncol(X)), estimate = 1,
  method = "sampling", ...)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Response times}

\item{v}{Censoring vector (0 = right-censored, 1 = exact time, 2 = left-censored, 3 = interval censored)}

\item{rcens}{If interval censored data, then the right-censoring time}

\item{mu0}{Prior mean for regression coefficients.}

\item{Sigma0}{Prior covariance for regression coefficients.}

\item{estimate}{Estimate the model (1) or just simulate data (0).}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan`
}
\description{
Bayesian proportional hazards exponential model
}
