% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survreg_stan.R
\name{survreg_ph_weibull}
\alias{survreg_ph_weibull}
\title{Bayesian proportional hazards weibull model}
\usage{
survreg_ph_weibull(X, y, v, rcens = rep(NA, length(y)), sigma0 = 1,
  gamma0 = 1/1000, ...)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Response times}

\item{v}{Censoring vector (0 = right-censored, 1 = exact time, 2 = left-censored, 3 = interval censored)}

\item{rcens}{If interval censored data, then the right-censoring time}

\item{...}{Other arguments}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian proportional hazards weibull model
}
